### 2025-10-07 #9168 @editorialbot
[PRE REVIEW]: PyTurbo_SF: An Adaptive Bootstrap Framework for Efficient Structure Function Analysis in Turbulent Flows
**Submitting author:** @aayouche (<a href="http://orcid.org/0009-0001-9075-5911">ADAM AYOUCHE</a>)
**Repository:** https://github.com/aayouche/pyturbo_sf
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.4
