### 2025-10-07 958f421
Update paper.md
Add Orcid to Adam Ayouche
