### 2025-10-12 4a05173
Update structure_functions.rst
Correct the structure functions definitions
