+++ docs/source/structure_functions.rst
@@ -10,9 +10,9 @@ Structure functions are fundamental tools in turbulence analysis that measure th
 
 .. math::
 
-   S_n(r) = \langle |\phi(\mathbf{x} + \mathbf{r}) - \phi(\mathbf{x})|^n \rangle_{\mathbf{x}}
+   S_n(r) = \langle |\phi(\vec{x} + \vec{r}) - \phi(\vec{x})|^n \rangle_{\vec{x}}
 
-where :math:`\phi` represents the field variable, :math:`\mathbf{r}` is the separation vector, and :math:`\langle \cdot \rangle_{\mathbf{x}}` denotes spatial averaging.
+where :math:`\phi` represents the field variable, :math:`\vec{r}` is the separation vector, and :math:`\langle \cdot \rangle_{\vec{x}}` denotes spatial averaging.
 
 1D Structure Functions
 ----------------------
@@ -22,11 +22,11 @@ For 1D data, PyTurbo_SF supports several types of structure functions:
 Scalar Structure Function
 ~~~~~~~~~~~~~~~~~~~~~~~~~
 
-For a scalar field :math:`\phi(x)`:
+For a scalar field :math:`f(x)`:
 
 .. math::
 
-   S_{\phi,n}(r) = \langle |\phi(x + r) - \phi(x)|^n \rangle_{x}
+   S_n(r) = \langle |f(x + r) - f(x)|^n \rangle_{x}
 
 where :math:`r` is the separation distance.
 
@@ -35,11 +35,11 @@ where :math:`r` is the separation distance.
 Scalar-Scalar Structure Function  
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-For two scalar fields :math:`\phi_1(x)` and :math:`\phi_2(x)`:
+For two scalar fields :math:`f(x)` and :math:`g(x)`:
 
 .. math::
 
-   S_{\phi_1\phi_2,n}(r) = \langle |(\phi_1(x + r) - \phi_1(x))(\phi_2(x + r) - \phi_2(x))|^n \rangle_{x}
+   S_{n,k}(r) = \langle |f(x + r) - f(x)|^n \cdot |g(x + r) - g(x)|^k \rangle_{x}
 
 **Usage**: ``fun='scalar_scalar'``
 
@@ -51,13 +51,13 @@ For 2D fields, PyTurbo_SF provides multiple types of structure functions:
 Longitudinal Structure Function
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-For velocity components :math:`\mathbf{u} = (u, v)`:
+For velocity components :math:`\vec{u} = (u, v)`:
 
 .. math::
 
-   S_{L,n}(r) = \langle |(\mathbf{u}(\mathbf{x} + \mathbf{r}) - \mathbf{u}(\mathbf{x})) \cdot \hat{\mathbf{r}}|^n \rangle_{\mathbf{x}}
+   S_{\parallel,n}(r) = \langle ((\vec{u}(\vec{x} + \vec{r}) - \vec{u}(\vec{x})) \cdot \frac{\vec{r}}{|\vec{r}|})^n \rangle_{\vec{x}}
 
-where :math:`\hat{\mathbf{r}} = \mathbf{r}/|\mathbf{r}|` is the unit vector in the direction of separation.
[…]
