### 2025-10-12 7730a00
Update paper.md
summarize the text
