+++ paper/paper.md
@@ -42,31 +42,32 @@ bibliography: paper.bib
 
 # Summary
 
-Structure functions are fundamental diagnostic tools in turbulence research that reveal energy cascade characteristics, intermittency properties, and scaling laws by quantifying spatial correlations between field differences at varying separation distances [@Frisch1995; @Pope2000]. PyTurbo_SF is a comprehensive Python package that tackles structure function analysis through an innovative adaptive bootstrap framework, enabling efficient and statistically rigorous calculations for 1D, 2D, and 3D turbulent datasets.
+Structure functions are fundamental diagnostic tools in turbulence research that quantify spatial correlations between field differences at varying separation distances, revealing energy cascade characteristics and scaling laws [@Frisch1995; @Pope2000]. PyTurbo_SF is a comprehensive Python package providing efficient, statistically rigorous structure function calculations for 1D, 2D, and 3D turbulent datasets through an innovative adaptive bootstrap framework.
 
-The package addresses critical computational bottlenecks in modern turbulence analysis by implementing power-of-2 spacing (sampling indices) strategies, adaptive convergence monitoring, and memory-optimized algorithms that automatically scale with dataset characteristics. PyTurbo_SF provides unprecedented statistical rigor through bootstrap resampling, delivering robust uncertainty quantification that has been missing from existing structure function tools. The software seamlessly processes xarray Datasets and supports diverse structure function types including longitudinal, transverse, scalar, advective, and energy flux functions recently developed for analyzing geophysical energy transfers [@Pearson2021; @Pearson2024].
+The package addresses computational bottlenecks through power-of-2 spacing strategies, adaptive convergence monitoring, and memory-optimized algorithms. PyTurbo_SF delivers robust uncertainty quantification through bootstrap resampling and supports diverse structure function types including longitudinal, transverse, scalar, advective, and energy flux functions [@Pearson2021; @Pearson2024].
 
-Applications span from oceanographic mooring time series and satellite altimetry measurements to high-resolution numerical simulations, enabling researchers to apply consistent, statistically sound methodology across scales from laboratory turbulence to planetary circulation systems.
+Applications span oceanographic time series and satellite measurements to high-resolution simulations, enabling consistent methodology across scales from laboratory to planetary systems.
 
 # Statement of need
 
-Contemporary turbulence research increasingly relies on massive, high-resolution datasets from advanced observational platforms (satellite missions, autonomous platforms) and computational simulations (large eddy simulations, direct numerical simulations). Traditional structure function calculations face severe limitations: computational intractability for large 2D/3D datasets, absence of rigorous uncertainty quantification, manual parameter tuning requirements, and limited function types that restrict scientific insights.
+Contemporary turbulence research relies on massive datasets from satellite missions, autonomous platforms, and high-resolution simulations. Traditional structure function calculations face severe limitations: computational intractability for large datasets, absence of uncertainty quantification, manual parameter tuning, and limited function types.
 
-Existing software tools address only subsets of these challenges. fastSF provides parallelized C++ implementations but lacks advanced function types and uncertainty quantification [@Sadhukhan2021]. MATLAB-based toolkits offer specific analyses but are environmentally limited and lack comprehensive statistical frameworks [@Fuchs2022]. Alternative approaches like coarse-graining (FlowSieve) address different aspects of turbulence analysis but cannot provide the detailed scale-by-scale information that structure functions uniquely deliver [@Storer2023].
[…]
