### 2025-10-12 0047494
Update paper.bib
update some references
