+++ paper/paper.bib
@@ -106,13 +106,6 @@
   doi={10.5334/jors.148}
 }
 
-@misc{Joblib2020,
-  title={Joblib: running {Python} functions as pipeline jobs},
-  author={{Joblib Development Team}},
-  year={2020},
-  url={https://joblib.readthedocs.io/},
-  note={Version 1.0.0}
-}
 
 @article{Morrow2018,
   title={Fine-scale altimetry and the future {SWOT} mission},
@@ -124,12 +117,15 @@
   doi={10.17125/gov2018.ch08}
 }
 
-@book{Monin1975,
-  title={Statistical Fluid Mechanics, Volume {II}: Mechanics of Turbulence},
-  author={Monin, Andrei S. and Yaglom, Akiva M.},
-  year={1975},
-  publisher={MIT Press},
-  address={Cambridge, MA}
+@article{Monin1975, 
+  title={Statistical Fluid Mechanics: Mechanics of Turbulence. Vol 2. A. S. Monin and A. M. Yaglom. MIT Press, Cambridge, Massachusetts. 1975. 874 pp. £25.00.}, 
+  volume={80}, 
+  DOI={10.1017/S000192400003342X}, 
+  number={781}, 
+  journal={The Aeronautical Journal}, 
+  author={Paterson, A. R.}, 
+  year={1976}, 
+  pages={44–44}
 }
 
 @article{Balwada2022,
