### 2025-10-26 #9168 @aayouche
@kthyng

I summarize the readme on the github page and cleaned some folders (delete some few unnecessary ones)

Please take your time - Thanks
