### 2025-12-16 094fb7d
v1.0.5: Documentation and example updates
- Add Bessel energy flux formulae and derivation
- Add conditional structure function documentation
- Improve function reference tables (1D/2D/3D)
- Fix sidebar contrast in documentation CSS
- Update example notebooks
