+++ src/pyturbo_sf/__init__.py
@@ -1,6 +1,6 @@
 # Import main functions to make them available at package level
 from .one_dimensional import bin_sf_1d
-from .two_dimensional import bin_sf_2d, get_isotropic_sf_2d
+from .two_dimensional import bin_sf_2d, get_isotropic_sf_2d, get_energy_flux_2d
 from .three_dimensional import bin_sf_3d, get_isotropic_sf_3d
 
-__version__ = "1.0.3"
+__version__ = "1.0.5"
