### 2025-12-16 62eac93
Fix ReadTheDocs build: update Python to 3.11
