+++ docs/requirements.txt
@@ -1,28 +1,30 @@
 # Documentation dependencies
-sphinx>=4.0
-sphinx-rtd-theme>=1.0
-sphinx-autoapi>=2.0
-myst-parser>=0.18
-nbsphinx>=0.8
-ipython>=7.0
-jupyter>=1.0
+sphinx>=7.0
+sphinx-rtd-theme>=2.0
+myst-parser>=2.0
+nbsphinx>=0.9
+ipython>=8.0
 pandoc>=2.0
 
 # Scientific computing dependencies for examples
-numpy>=1.20
-xarray>=0.19
-matplotlib>=3.5
-scipy>=1.7
-pandas>=1.3
+numpy>=1.24
+xarray>=2023.0
+matplotlib>=3.7
+scipy>=1.10
+pandas>=2.0
 
 # For interactive plots in docs
 plotly>=5.0
-bokeh>=2.4
+bokeh>=3.0
 
 # Notebook execution
-nbconvert>=6.0
+nbconvert>=7.0
 ipykernel>=6.0
 
 # Additional formatting
 sphinx-copybutton>=0.5
-sphinx-design>=0.3
+sphinx-design>=0.5
+
+# Core package dependencies (needed for autodoc)
+bottleneck>=1.3
+joblib>=1.3
