+++ docs/source/pyturbo_sf.rst
@@ -1,19 +1,93 @@
 pyturbo\_sf package
 ===================
 
-.. automodule:: pyturbo_sf
+Submodules
+----------
+
+pyturbo\_sf.bessel\_tools module
+--------------------------------
+
+.. automodule:: pyturbo_sf.bessel_tools
    :members:
    :undoc-members:
    :show-inheritance:
 
-Submodules
-----------
+pyturbo\_sf.binning\_tools module
+---------------------------------
 
-.. toctree::
-   :maxdepth: 4
+.. automodule:: pyturbo_sf.binning_tools
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
-   pyturbo_sf.core
-   pyturbo_sf.one_dimensional
-   pyturbo_sf.three_dimensional
-   pyturbo_sf.two_dimensional
-   pyturbo_sf.utils
+pyturbo\_sf.bootstrapping\_tools module
+---------------------------------------
+
+.. automodule:: pyturbo_sf.bootstrapping_tools
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyturbo\_sf.core module
+-----------------------
+
+.. automodule:: pyturbo_sf.core
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyturbo\_sf.isotropy\_tools module
+----------------------------------
+
+.. automodule:: pyturbo_sf.isotropy_tools
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyturbo\_sf.one\_dimensional module
+-----------------------------------
+
+.. automodule:: pyturbo_sf.one_dimensional
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyturbo\_sf.structure\_functions module
+---------------------------------------
+
+.. automodule:: pyturbo_sf.structure_functions
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyturbo\_sf.three\_dimensional module
+-------------------------------------
+
+.. automodule:: pyturbo_sf.three_dimensional
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyturbo\_sf.two\_dimensional module
+-----------------------------------
+
+.. automodule:: pyturbo_sf.two_dimensional
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyturbo\_sf.utils module
+------------------------
+
+.. automodule:: pyturbo_sf.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyturbo_sf
+   :members:
+   :undoc-members:
+   :show-inheritance:
