+++ docs/source/modules.rst
@@ -1,5 +1,7 @@
-pyturbo_sf
-==========
+API Reference
+=============
+
+This section contains the complete API reference for PyTurbo_SF.
 
 .. toctree::
    :maxdepth: 4
