+++ docs/source/pyturbo_sf.binning_tools.rst
@@ -1,0 +1,7 @@
+pyturbo\_sf.binning\_tools module
+=================================
+
+.. automodule:: pyturbo_sf.binning_tools
+   :members:
+   :undoc-members:
+   :show-inheritance:
