+++ docs/source/pyturbo_sf.rst
@@ -1,93 +1,52 @@
 pyturbo\_sf package
 ===================
 
-Submodules
-----------
-
-pyturbo\_sf.bessel\_tools module
---------------------------------
-
-.. automodule:: pyturbo_sf.bessel_tools
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyturbo\_sf.binning\_tools module
----------------------------------
-
-.. automodule:: pyturbo_sf.binning_tools
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyturbo\_sf.bootstrapping\_tools module
----------------------------------------
-
-.. automodule:: pyturbo_sf.bootstrapping_tools
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyturbo\_sf.core module
------------------------
-
-.. automodule:: pyturbo_sf.core
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyturbo\_sf.isotropy\_tools module
-----------------------------------
-
-.. automodule:: pyturbo_sf.isotropy_tools
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyturbo\_sf.one\_dimensional module
------------------------------------
-
-.. automodule:: pyturbo_sf.one_dimensional
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyturbo\_sf.structure\_functions module
----------------------------------------
-
-.. automodule:: pyturbo_sf.structure_functions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyturbo\_sf.three\_dimensional module
--------------------------------------
-
-.. automodule:: pyturbo_sf.three_dimensional
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyturbo\_sf.two\_dimensional module
------------------------------------
-
-.. automodule:: pyturbo_sf.two_dimensional
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyturbo\_sf.utils module
-------------------------
-
-.. automodule:: pyturbo_sf.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
 .. automodule:: pyturbo_sf
    :members:
    :undoc-members:
    :show-inheritance:
+
+Main API Functions
+------------------
+
+The following functions are available at the package level:
+
+.. currentmodule:: pyturbo_sf
+
+1D Structure Functions
+~~~~~~~~~~~~~~~~~~~~~~
+
+.. autofunction:: bin_sf_1d
+
+2D Structure Functions
+~~~~~~~~~~~~~~~~~~~~~~
+
+.. autofunction:: bin_sf_2d
+
+.. autofunction:: get_isotropic_sf_2d
+
+.. autofunction:: get_energy_flux_2d
+
+3D Structure Functions
+~~~~~~~~~~~~~~~~~~~~~~
+
+.. autofunction:: bin_sf_3d
+
+.. autofunction:: get_isotropic_sf_3d
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 2
+
+   pyturbo_sf.one_dimensional
+   pyturbo_sf.two_dimensional
+   pyturbo_sf.three_dimensional
+   pyturbo_sf.core
+   pyturbo_sf.utils
+   pyturbo_sf.structure_functions
+   pyturbo_sf.binning_tools
+   pyturbo_sf.bootstrapping_tools
+   pyturbo_sf.isotropy_tools
+   pyturbo_sf.bessel_tools
