+++ docs/source/pyturbo_sf.isotropy_tools.rst
@@ -1,0 +1,7 @@
+pyturbo\_sf.isotropy\_tools module
+==================================
+
+.. automodule:: pyturbo_sf.isotropy_tools
+   :members:
+   :undoc-members:
+   :show-inheritance:
