### 2025-12-16 3f624e7
Fix documentation: resolve merge conflicts, add all API modules
