+++ docs/source/pyturbo_sf.three_dimensional.rst
@@ -1,6 +1,8 @@
 pyturbo\_sf.three\_dimensional module
 =====================================
 
+3D structure function calculations including spherical isotropic averaging.
+
 .. automodule:: pyturbo_sf.three_dimensional
    :members:
    :undoc-members:
