+++ docs/source/pyturbo_sf.utils.rst
@@ -1,6 +1,8 @@
 pyturbo\_sf.utils module
 ========================
 
+Utility functions for variable mapping, confidence intervals, quality masks, and helper operations.
+
 .. automodule:: pyturbo_sf.utils
    :members:
    :undoc-members:
