+++ docs/source/pyturbo_sf.two_dimensional.rst
@@ -1,6 +1,8 @@
 pyturbo\_sf.two\_dimensional module
 ===================================
 
+2D structure function calculations including isotropic averaging and energy flux decomposition.
+
 .. automodule:: pyturbo_sf.two_dimensional
    :members:
    :undoc-members:
