+++ docs/source/pyturbo_sf.bootstrapping_tools.rst
@@ -1,6 +1,8 @@
 pyturbo\_sf.bootstrapping\_tools module
 =======================================
 
+Bootstrap resampling tools for uncertainty quantification and adaptive convergence.
+
 .. automodule:: pyturbo_sf.bootstrapping_tools
    :members:
    :undoc-members:
