+++ docs/source/pyturbo_sf.core.rst
@@ -1,6 +1,8 @@
 pyturbo\_sf.core module
 =======================
 
+Core utilities for dataset validation, bootstrap setup, and spacing calculations.
+
 .. automodule:: pyturbo_sf.core
    :members:
    :undoc-members:
