+++ docs/source/pyturbo_sf.bessel_tools.rst
@@ -1,6 +1,8 @@
 pyturbo\_sf.bessel\_tools module
 ================================
 
+Bessel transform tools for spectral energy flux decomposition using J1 transforms.
+
 .. automodule:: pyturbo_sf.bessel_tools
    :members:
    :undoc-members:
