+++ docs/source/pyturbo_sf.structure_functions_module.rst
@@ -1,0 +1,9 @@
+pyturbo\_sf.structure\_functions\_module module
+===============================================
+
+Core structure function computation routines for 1D, 2D, and 3D fields.
+
+.. automodule:: pyturbo_sf.structure_functions
+   :members:
+   :undoc-members:
+   :show-inheritance:
