+++ docs/source/structure_functions.rst
@@ -12,279 +12,150 @@ This document provides the mathematical foundations and API reference for struct
 
 ----
 
-<<<<<<< HEAD
 Mathematical Foundations
 ========================
 
-Standard Structure Functions
------------------------------
+Structure functions measure statistical properties of turbulent field increments across separation distances. The general form is:
 
-Structure functions measure statistical properties of turbulent field increments across separation distances.
-=======
+.. math::
+
    S_n(r) = \langle |\phi(\vec{x} + \vec{r}) - \phi(\vec{x})|^n \rangle_{\vec{x}}
 
 where :math:`\phi` represents the field variable, :math:`\vec{r}` is the separation vector, and :math:`\langle \cdot \rangle_{\vec{x}}` denotes spatial averaging.
 
+
 1D Structure Functions
 ----------------------
 
-For 1D data, PyTurbo_SF supports several types of structure functions:
+For 1D data (time series), PyTurbo_SF supports:
 
-Scalar Structure Function
-~~~~~~~~~~~~~~~~~~~~~~~~~
-
-For a scalar field :math:`f(x)`:
+**Scalar Structure Function** (``fun='scalar'``)
 
 .. math::
 
    S_n(r) = \langle |f(x + r) - f(x)|^n \rangle_{x}
 
-where :math:`r` is the separation distance.
-
-**Usage**: ``fun='scalar'``
-
-Scalar-Scalar Structure Function  
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-For two scalar fields :math:`f(x)` and :math:`g(x)`:
+**Scalar-Scalar Structure Function** (``fun='scalar_scalar'``)
 
 .. math::
 
    S_{n,k}(r) = \langle |f(x + r) - f(x)|^n \cdot |g(x + r) - g(x)|^k \rangle_{x}
 
-**Usage**: ``fun='scalar_scalar'``
+.. important::
+
+   **1D only supports** ``scalar`` and ``scalar_scalar`` function types.
+
 
 2D Structure Functions
 ----------------------
 
-For 2D fields, PyTurbo_SF provides multiple types of structure functions:
->>>>>>> a204517c41958803e48d06e78c5c6b50e8b2929d
+For 2D velocity fields :math:`\vec{u} = (u, v)`:
 
-Longitudinal Structure Function
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+**Longitudinal Structure Function** (``fun='longitudinal'``)
 
-<<<<<<< HEAD
-The longitudinal structure function projects velocity increments onto the separation vector:
+Projects velocity increments onto the separation vector:
 
 .. math::
 
-   D_{LL}(\mathbf{r}) = \left\langle \left[ \delta u_L(\mathbf{x}, \mathbf{r}) \right]^2 \right\rangle
+   D_{LL}(\mathbf{r}) = \left\langle \left[ \delta u_L \right]^n \right\rangle
 
-where the longitudinal velocity increment is:
+where :math:`\delta u_L = [\mathbf{u}(\mathbf{x} + \mathbf{r}) - \mathbf{u}(\mathbf{x})] \cdot \hat{\mathbf{r}}`
 
-.. math::
+**Transverse Structure Function** (``fun='transverse'``)
 
[…]
