+++ docs/source/pyturbo_sf.rst
@@ -34,6 +34,7 @@ The following functions are available at the package level:
 
 .. autofunction:: get_isotropic_sf_3d
 
+
 Submodules
 ----------
 
@@ -43,7 +44,6 @@ Submodules
    pyturbo_sf.one_dimensional
    pyturbo_sf.two_dimensional
    pyturbo_sf.three_dimensional
-   pyturbo_sf.structure_functions_module
    pyturbo_sf.core
    pyturbo_sf.utils
    pyturbo_sf.binning_tools
