### 2025-12-16 b0fda1f
v1.0.5: Fix documentation - add all API modules, fix tables, improve sidebar
