+++ docs/source/pyturbo_sf.structure_functions.rst
@@ -1,7 +1,0 @@
-pyturbo\_sf.structure\_functions module
-=======================================
-
-.. automodule:: pyturbo_sf.structure_functions
-   :members:
-   :undoc-members:
-   :show-inheritance:
