+++ docs/source/pyturbo_sf.isotropy_tools.rst
@@ -1,9 +1,0 @@
-pyturbo\_sf.isotropy\_tools module
-==================================
-
-Isotropy analysis tools for 2D (polar) and 3D (spherical) coordinate averaging.
-
-.. automodule:: pyturbo_sf.isotropy_tools
-   :members:
-   :undoc-members:
-   :show-inheritance:
