+++ docs/source/pyturbo_sf.binning_tools.rst
@@ -1,9 +1,0 @@
-pyturbo\_sf.binning\_tools module
-=================================
-
-Binning utilities including conditional structure function support with indicator functions.
-
-.. automodule:: pyturbo_sf.binning_tools
-   :members:
-   :undoc-members:
-   :show-inheritance:
