+++ docs/source/pyturbo_sf.core.rst
@@ -1,9 +1,0 @@
-pyturbo\_sf.core module
-=======================
-
-Core utilities for dataset validation, bootstrap setup, and spacing calculations.
-
-.. automodule:: pyturbo_sf.core
-   :members:
-   :undoc-members:
-   :show-inheritance:
