+++ docs/source/pyturbo_sf.three_dimensional.rst
@@ -1,9 +1,0 @@
-pyturbo\_sf.three\_dimensional module
-=====================================
-
-3D structure function calculations including spherical isotropic averaging.
-
-.. automodule:: pyturbo_sf.three_dimensional
-   :members:
-   :undoc-members:
-   :show-inheritance:
