+++ docs/source/pyturbo_sf.one_dimensional.rst
@@ -1,9 +1,0 @@
-pyturbo\_sf.one\_dimensional module
-===================================
-
-1D structure function calculations for time series and single-dimension data.
-
-.. automodule:: pyturbo_sf.one_dimensional
-   :members:
-   :undoc-members:
-   :show-inheritance:

[…]
