### 2025-12-16 64da288
Fix documentation: resolve merge conflicts, add all API modules
