### 2025-12-16 e1ca57c
Fix documentation: resolve merge conflicts, add all API modules
