### 2025-12-16 6c1be81
Fix documentation: resolve merge conflicts, add all API modules
