+++ docs/source/conf.py
@@ -4,7 +4,7 @@
 
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../src'))  # Point directly to your package
+sys.path.insert(0, os.path.abspath('../../src'))  # Point directly to your package
 
 # -- Project information -----------------------------------------------------
 project = 'PyTurbo_SF'
