### 2025-12-16 e2e7b68
Regenerate API documentation
