+++ docs/source/pyturbo_sf.rst
@@ -10,6 +10,7 @@ pyturbo\_sf.bessel\_tools module
 .. automodule:: pyturbo_sf.bessel_tools
    :members:
    :undoc-members:
+   :private-members:     
    :show-inheritance:
 
 pyturbo\_sf.binning\_tools module
@@ -18,6 +19,7 @@ pyturbo\_sf.binning\_tools module
 .. automodule:: pyturbo_sf.binning_tools
    :members:
    :undoc-members:
+   :private-members:     
    :show-inheritance:
 
 pyturbo\_sf.bootstrapping\_tools module
@@ -26,6 +28,7 @@ pyturbo\_sf.bootstrapping\_tools module
 .. automodule:: pyturbo_sf.bootstrapping_tools
    :members:
    :undoc-members:
+   :private-members:     
    :show-inheritance:
 
 pyturbo\_sf.core module
@@ -34,6 +37,7 @@ pyturbo\_sf.core module
 .. automodule:: pyturbo_sf.core
    :members:
    :undoc-members:
+   :private-members:     
    :show-inheritance:
 
 pyturbo\_sf.isotropy\_tools module
@@ -42,6 +46,7 @@ pyturbo\_sf.isotropy\_tools module
 .. automodule:: pyturbo_sf.isotropy_tools
    :members:
    :undoc-members:
+   :private-members:     
    :show-inheritance:
 
 pyturbo\_sf.one\_dimensional module
@@ -50,6 +55,7 @@ pyturbo\_sf.one\_dimensional module
 .. automodule:: pyturbo_sf.one_dimensional
    :members:
    :undoc-members:
+   :private-members:     
    :show-inheritance:
 
 pyturbo\_sf.structure\_functions module
@@ -58,6 +64,7 @@ pyturbo\_sf.structure\_functions module
 .. automodule:: pyturbo_sf.structure_functions
    :members:
    :undoc-members:
+   :private-members:     
    :show-inheritance:
 
 pyturbo\_sf.three\_dimensional module
@@ -66,6 +73,7 @@ pyturbo\_sf.three\_dimensional module
 .. automodule:: pyturbo_sf.three_dimensional
    :members:
    :undoc-members:
+   :private-members:     
    :show-inheritance:
 
 pyturbo\_sf.two\_dimensional module
@@ -74,6 +82,7 @@ pyturbo\_sf.two\_dimensional module
 .. automodule:: pyturbo_sf.two_dimensional
    :members:
    :undoc-members:
+   :private-members:     
    :show-inheritance:
 
 pyturbo\_sf.utils module
@@ -82,6 +91,7 @@ pyturbo\_sf.utils module
 .. automodule:: pyturbo_sf.utils
    :members:
    :undoc-members:
+   :private-members:     
    :show-inheritance:
 
 Module contents
@@ -90,4 +100,5 @@ Module contents
 .. automodule:: pyturbo_sf
    :members:
    :undoc-members:
+   :private-members:     
    :show-inheritance:
