+++ docs/source/examples.rst
@@ -4,27 +4,29 @@ Examples and Tutorials
 This section provides comprehensive examples demonstrating PyTurbo_SF capabilities for real-world oceanographic and atmospheric applications.
 
 Jupyter Notebook Examples
---------------------------
+-------------------------
 
 **Complete Application Examples:**
 
 - :doc:`examples/example_1D_mooring` - Structure function analysis of time series from oceanographic moorings
 - :doc:`examples/example_2D_SWOT` - Satellite altimetry data analysis using 2D structure functions  
 - :doc:`examples/example_2D_CROCO` - Regional ocean modeling analysis with CROCO simulation data
-- :doc:`examples/example_2D_CROCO_Bessel` - Regional ocean modeling analysis with CROCO simulation data For Bessel Energy Flux Decomposition
-- :doc:`examples/example_2D_MITGCM_Conditional` - Regional ocean modeling analysis with LLC4320 simulation data For Conditional Structure Function Based on Sea Ice Concentration
+- :doc:`examples/example_2D_CROCO_Bessel` - Regional ocean modeling analysis with CROCO simulation data for Bessel Energy Flux Decomposition
+- :doc:`examples/example_2D_MITGCM_Conditional` - Regional ocean modeling analysis with LLC4320 simulation data for Conditional Structure Function Based on Sea Ice Concentration
 - :doc:`examples/example_3D_DYCOMS` - Atmospheric boundary layer turbulence analysis from DYCOMS LES
 
 .. toctree::
+   :maxdepth: 1
    :hidden:
 
-   examples/1d_time_mooring
-   examples/2d_swot
-   examples/2d_croco
-   examples/3d_dycoms
+   examples/example_1D_mooring
+   examples/example_2D_SWOT
+   examples/example_2D_CROCO
+   examples/example_2D_CROCO_Bessel
+   examples/example_2D_MITGCM_Conditional
+   examples/example_3D_DYCOMS
 
 .. note::
    **Additional Examples Available**
    
    More examples and use cases are available in the `GitHub repository examples directory <https://github.com/aayouche/pyturbo_sf/tree/main/docs/source/examples>`_. These include specialized applications for different scientific domains and advanced analysis techniques.
-
