### 2025-12-17 39beeea
Delete src/pyturbo_sf.egg-info directory
