+++ src/pyturbo_sf.egg-info/SOURCES.txt
@@ -1,30 +1,0 @@
-LICENSE
-README.md
-pyproject.toml
-src/pyturbo_sf/__init__.py
-src/pyturbo_sf/bessel_tools.py
-src/pyturbo_sf/binning_tools.py
-src/pyturbo_sf/bootstrapping_tools.py
-src/pyturbo_sf/core.py
-src/pyturbo_sf/isotropy_tools.py
-src/pyturbo_sf/one_dimensional.py
-src/pyturbo_sf/structure_functions.py
-src/pyturbo_sf/three_dimensional.py
-src/pyturbo_sf/two_dimensional.py
-src/pyturbo_sf/utils.py
-src/pyturbo_sf.egg-info/PKG-INFO
-src/pyturbo_sf.egg-info/SOURCES.txt
-src/pyturbo_sf.egg-info/dependency_links.txt
-src/pyturbo_sf.egg-info/requires.txt
-src/pyturbo_sf.egg-info/top_level.txt
-tests/test_bessel_tools.py
-tests/test_binning_tools.py
-tests/test_bootstrapping_tools.py
-tests/test_core.py
-tests/test_isotropy_tools.py
-tests/test_main.py
-tests/test_one_dimensional.py
-tests/test_structure_functions.py
-tests/test_three_dimensional.py
-tests/test_two_dimensional.py
-tests/test_utils.py
