### 2026-01-01 40a88ba
v1.0.6: Statistical improvements and bug fixes
Code changes:
- Delete ci_method parameter
- Add seed parameter for bootstrap reproducibility
- Fix point_counts to count pairs (x, x+r) correctly
- Fix weighting to match pair counts
- Consistent weighted standard error for all methods
- Set add_to_counts=True in adaptive bootstrap sampling
- Add diagnostics

Documentation:
- Update examples
- Update paper (FluidSF comparison)
- Update pyproject.toml dependencies
