+++ paper/paper.bib
@@ -162,3 +162,14 @@
   publisher={Nature Publishing Group},
   doi={10.1038/nphys4227}
 }
+
+@article{Wagner2025,
+  title={{FluidSF}: A {Python} package for calculating turbulent flow statistics},
+  author={Wagner, Cassidy M. and Lee, Ara and Pearson, Brodie},
+  journal={Journal of Open Source Software},
+  volume={10},
+  number={114},
+  pages={7873},
+  year={2025},
+  doi={10.21105/joss.07873}
+}
