+++ src/pyturbo_sf/bootstrapping_tools.py
@@ -34,110 +34,120 @@ from .bessel_tools import (
 from scipy.special import jv  # Bessel function of the first kind
 
 
-def _weighted_percentile(values, weights, percentile):
-    """
-    Compute weighted percentile.
-    
-    Parameters
-    ----------
-    values : array
-        Data values
-    weights : array
-        Weights for each value
-    percentile : float
-        Percentile to compute (0-100)
-        
-    Returns
-    -------
-    float
-        Weighted percentile value
-    """
-    # Sort by values
-    sorted_indices = np.argsort(values)
-    sorted_values = values[sorted_indices]
-    sorted_weights = weights[sorted_indices]
-    
-    # Compute cumulative weights
-    cumulative_weights = np.cumsum(sorted_weights)
-    total_weight = cumulative_weights[-1]
-    
-    # Normalize to 0-100 scale
-    cumulative_percentiles = 100 * (cumulative_weights - 0.5 * sorted_weights) / total_weight
-    
-    # Interpolate to find the value at the desired percentile
-    return np.interp(percentile, cumulative_percentiles, sorted_values)
+##################################################
+# BOOTSTRAP STATISTICS
+##################################################
 
 
-def _compute_weighted_bootstrap_stats(bootstrap_samples, confidence_level=0.95, ci_method='percentile'):
+def _compute_weighted_bootstrap_stats(bootstrap_samples, confidence_level=0.95):
     """
-    Compute weighted mean, std, and confidence intervals from bootstrap samples.
+    Compute bootstrap statistics with proper effective sample size correction.
     
     Parameters
     ----------
     bootstrap_samples : list of dict
-        Each dict contains 'mean' and 'weight'
+        Each dict contains 'mean' and 'weight' (number of points in that bootstrap)
     confidence_level : float
-        Confidence level for intervals
-    ci_method : str
-        'percentile' or 'standard'
+        Confidence level for intervals (default: 0.95)
         
     Returns
     -------
-    weighted_mean : float
-        Weighted mean of bootstrap means
-    weighted_std : float
-        Weighted standard deviation
+    theta_hat : float
+        Point estimate (weighted mean of bootstrap means)
+    std_error : float
+        Bootstrap standard error with effective sample size correction
     ci_lower : float
-        Lower confidence interval
+        Lower confidence interval bound (theta_hat - z * SE)
     ci_upper : float
-        Upper confidence interval
+        Upper confidence interval bound (theta_hat + z * SE)
+        
+    Notes
+    -----
+    The standard error is computed using effective sample size:
+    
+    1. n_eff = (sum(w))^2 / sum(w^2)
+    2. var_corrected = var_weighted * n_eff / (n_eff - 1)  [Bessel correction]
+    3. SE = sqrt(var_corrected / n_eff)
+    
+    This properly accounts for:
+    - Unequal weights in bootstrap samples
+    - Bias correction (Bessel's correction)
+    - Variance of the mean (not variance of data)
     """
+    from scipy import stats
+    
     boot_means = np.array([s['mean'] for s in bootstrap_samples])
-    boot_weights = np.array([s['weight'] for s in bootstrap_samples])
+    boot_weights = np.array([s['weight'] for s in bootstrap_samples], dtype=np.float64)
     
-    # Weighted mean
-    weighted_mean = np.average(boot_means, weights=boot_weights)
+    # Handle edge cases
+    if len(boot_means) == 0:
+        return np.nan, np.nan, np.nan, np.nan
     
-    # Weighted variance and std
-    weighted_var = np.average((boot_means - weighted_mean)**2, weights=boot_weights)
-    weighted_std = np.sqrt(weighted_var)
+    if len(boot_means) == 1:
[…]
