### 2026-01-09 40390c3
Fix dimension order detection in validate_dataset functions
