### 2026-01-09 31a93aa
Fix dimension order detection in validate_dataset functions
