+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pyturbo_sf"
-version = "1.0.6"
+version = "1.0.7"
 description = "Python Turbulence Unleashed: Rapid Binning Operator"
 readme = "README.md"
 authors = [
