### 2026-01-12 #9168 @Sbozzolo
> [@Sbozzolo](https://github.com/Sbozzolo) I have updated the paper (add comparison with fluidsf based on [@MaceKuailv](https://github.com/MaceKuailv) comment) and the package (add some diagnostics see docs and updated some bugs), version 1.0.7 is the latest stable - thanks and thank you [@NoraLoose](https://github.com/NoraLoose) for accepting to review the package - wish you all happy new year !

Thank you! I am looking for the second reviewer, and I reached out to some people over email. Sorry for the delay.
