### 2026-01-22 #9876 @editorialbot
[REVIEW]: PyTurbo_SF: An Adaptive Bootstrap Framework for Efficient Structure Function Analysis in Turbulent Flows
**Submitting author:** @aayouche (<a href="http://orcid.org/0009-0001-9075-5911">ADAM AYOUCHE</a>)
**Repository:** https://github.com/aayouche/pyturbo_sf
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.8
