### 2026-02-03 #9876 @einaraz
Hi @aayouche and @Sbozzolo:  I have gone through the entire package and I think the project is very well done and well documented.  A few comments:

1. The package name and initial logo suggest that it is intended for turbulence analysis, which is a very broad field. However, it is designed for structure function analysis, a specific type of analysis within turbulence. I wonder if this should be clarified earlier, including the description in the logo.
2. Paper: one major source of turbulent data comes from field experiments using eddy-covariance systems, which is not mentioned. This package would be very useful for this community and an example could be easily added in the tutorial if the authors wanted to explore that area.
3. Figures 1 and 2: Please define all variables in the figure captions. It would also help to include a short paragraph in the main text explaining the figures and specifying which functions were used to generate them and what raw data served as input.
4. Installation: I successfully installed the package using pip install pyturbo_sf (tested on Python 3.9). However, installation from the repository appears to require Python 3.11 or higher. Additionally, the pyproject.toml and requirements.txt files inside the docs/ directory seem mismatched.
5. Documentation: The documentation is very detailed, and I was able to run several of the provided examples successfully.
