### 2026-03-04 #9876 @NoraLoose
@aayouche I'm currently reading through the paper. It's overall great. Here are a few comments:

* **Overarching comment for figures:** It would greatly help readers if the figure captions were more descriptive and help guide the reader through the panels and data.

    - **Figure 1:** The caption could be a bit more descriptive. For example, you could explain that the upper panel shows zonal velocity from SWOT along the satellite track (maybe with citation for SWOT to point non-expert readers to a reference), and the lower panel shows the structure function analysis. Maybe you could clarify what “longitudinal” means. Does it mean the separation vector points along longitude? In the upper panel, is fixed latitude along a fixed y? Adding an arrow in the upper panel to show the separation vector direction could help. Also, do you average over the entire SWOT path? You mention “bootstrap error bars,” but I don’t see any in the lower panel. It could also help to include a reference line for the expected r^{2/3} scaling.

  - **Figure 2:** Maybe add a citation for CROCO. And you could describe the panels in the caption so readers can follow along. Even if you don’t explain every line on the right (different structure function methods?), you could say something like: “The figure shows structure functions computed via four approaches from velocity and vorticity data” (if that's correct). Does the bottom-right panel show uncertainty? It would also help to mention that the figure compares “third-order & advective” versus “traditional second-order” approaches.

  - **Figure 3:** It would help to clarify the axes. What exactly do x and y represent? If the figure highlights anisotropic features, you could state that explicitly. :)

  - **Figure 4:** Maybe include a citation for DYCOMS. And what exactly does the figure show?

* Line 67: Abbreviation SF for structure function has not been introduced.
* Maybe mention FluidSF already in the Statement of need when you discuss other related tools. (You could say that you will discuss differences between PyTurbo_SF and FluidSF further below.)
