### 2026-03-04 aayouche/pyturbo_sf#4 @NoraLoose
Badges in README aren't rendering
The badges in the README aren't rendering correctly I think.

This issue is part of the JOSS review over [here](https://github.com/openjournals/joss-reviews/issues/9876).
