### 2026-03-04 #9876 @NoraLoose
@aayouche, in order for me to evaluate/check this box, could you briefly explain how your co-authors contributed to the software? Thanks!

> Contribution and authorship: Has the submitting author (@aayouche) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
