### 2026-03-04 #9876 @NoraLoose
> As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual.

@aayouche, in order to check this box, could you briefly clarify whether this has taken more than three months of your time? From the GitHub commit history, it looks like it may be less.
