### 2026-03-04 aayouche/pyturbo_sf#5 @NoraLoose
List in documentation doesn't render correctly
This issue is part of the JOSS review over on https://github.com/openjournals/joss-reviews/issues/9876.
