### 2026-03-04 #9876 @aayouche
> [@aayouche](https://github.com/aayouche), in order for me to evaluate/check this box, could you briefly explain how your co-authors contributed to the software? Thanks!
>
> > Contribution and authorship: Has the submitting author ([@aayouche](https://github.com/aayouche)) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

Hi @NoraLoose , thank you for the question! Here is a brief summary of each co-author's contributions to the software:

- Baylor Fox-Kemper (supervisor): provided overall supervision, contributed to the methodology behind the bootstrapping and benchmarking approaches, and guided the theoretical framework underlying the package design, in addition to editing the paper.

- Momme Hell: led the benchmarking effort, carried out extensive testing of the code, and served as an overall checker of the software's correctness and robustness.

- Cassidy Wagner & Brodie Pearson: contributed the methodology for the advective structure functions and the Bessel-transform-based energy flux computations, and reviewed the paper.

Happy to clarify further if needed!
