### 2026-03-04 aayouche/pyturbo_sf#6 @NoraLoose
Reduce argument repetition in example notebooks
In your example https://pyturbo-sf.readthedocs.io/en/latest/examples/example_1D_mooring.html, you could avoid repetition of shared arguments by defining a common `kwargs` dictionary, e.g.:

```
kwargs = {
    "ds": ds_new,
    "variable_names": ["u"],
    "order": 2,
    "bins": bins,
    "fun": 'scalar',
    "initial_nbootstrap": 100,
    "max_nbootstrap": 200,
    "step_nbootstrap": 50,
    "convergence_eps": 0.02,
    "n_jobs": -1,
    "backend": 'loky',
    "seed": 42
}
```
Then the following cells become much more compact and the differences between runs are immediately visible (e.g., only `bootsize` changes):

```
%%time

bin_time = np.logspace(np.log10(dt_min), np.log10(653819.0), 10)
bins = {
    'time': bin_time,
}
iso_sf_lon = psf.bin_sf_1d(
    bootsize={'time':10898},
    **kwargs
)
```

This pattern might also be helpful in the other example notebooks. Of course, this is optional, but I think it would make the notebooks more compact and easier to read.

I also noticed that in a few places the `bootsize` stated in the caption does not match the value actually used in the code. It would be good to double-check those for consistency.

This issue is part of the JOSS review over on https://github.com/openjournals/joss-reviews/issues/9876.
