### 2026-03-04 #9876 @aayouche
> > As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual.
>
> [@aayouche](https://github.com/aayouche), in order to check this box, could you briefly clarify whether this has taken more than three months of your time? From the GitHub commit history, it looks like it may be less.

Hi @NoraLoose , great question — and I want to clarify this, as the GitHub commit history alone can be misleading.

The package grew out of in-house code that I had been developing and actively using in my research for roughly a year before the repository was made public. The early development happened privately, in the context of ongoing projects on turbulence structure function analysis, and the codebase was progressively refined, tested, and extended over that period before I decided to formalize it into a publishable open-source package.

I actually flagged this context during the pre-review stage with one of the editors, noting that the public commit history would not fully reflect the true development timeline. The version that was eventually published on GitHub represents the culmination of that longer, ongoing effort — not a project started from scratch a few months ago.

I hope this clarifies things, and I'm happy to provide any additional context if needed!
