### 2026-03-04 aayouche/pyturbo_sf#7 @NoraLoose
Is pyturbo the same as pyturbo_sf?
In the docs, I see mixed usage of `pyturbo` and `pyturbo_sf`. Are these the same thing? If so, would it be better to stay consistent with the naming to avoid confusion?

This issue is part of the JOSS review over on https://github.com/openjournals/joss-reviews/issues/9876.
