### 2026-03-04 aayouche/pyturbo_sf#8 @NoraLoose
Benchmarking results
In the JOSS paper, you write

> Benchmark testing demonstrates O(NM log N log M) complexity for 2D data, enabling analysis of datasets with millions of grid points.

Are these results shown anywhere? Just asking because I'm evaluating this point

> Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

as part of the JOSS review over here: https://github.com/openjournals/joss-reviews/issues/9876.
