### 2026-03-04 aayouche/pyturbo_sf#8 @aayouche
please see here

https://pyturbo-sf.readthedocs.io/en/latest/benchmark/benchmark_analysis.html#Plot-Benchmark-Results
