### 2026-03-04 aayouche/pyturbo_sf#9 @NoraLoose
Illustrate impact of bootstrap block size in 1D example
In the example notebook https://pyturbo-sf.readthedocs.io/en/latest/examples/example_1D_mooring.html you test different bootstrap block sizes. Is this mainly for performance comparison? It would also be very interesting to explicitly show how the different block sizes affect the results, both the structure function itself and the associated uncertainty estimates. This could also serve as a concrete illustration of what you discuss here: https://pyturbo-sf.readthedocs.io/en/latest/optimal_bootsize.html (if I understand correctly). For example, you could add a final plot in the 1D notebook comparing the results obtained with different bootsize choices.

Of course, this is optional, but I think it would be very helpful for users to see the practical impact of the bootstrap block size.

This comment is part of the JOSS review at: https://github.com/openjournals/joss-reviews/issues/9876
