### 2026-03-04 aayouche/pyturbo_sf#10 @NoraLoose
Interpretation of figures/results
I think it would be very helpful for users if the example notebooks included a short sentence or two interpreting what the plots actually show.

For example:

* In the 1D mooring example https://pyturbo-sf.readthedocs.io/en/latest/examples/example_1D_mooring.html the motivation states: “The 1D structure function might reveal some time scales at which turbulent energy is transferred in the ocean.” It would be great to briefly return to this point after the figure and explain what the plotted result tells us physically.

* Similarly, in the 2D SWOT example https://pyturbo-sf.readthedocs.io/en/latest/examples/example_2D_SWOT.html, it would be helpful to add one or two sentences under each figure providing a broad physical interpretation.

* Similarly, for the other notebooks

This is of course optional, but I think a bit more interpretation would make the examples more instructive, especially for users who are less familiar with structure functions. :)

This issue is part of the JOSS review over on https://github.com/openjournals/joss-reviews/issues/9876.
