### 2026-03-04 aayouche/pyturbo_sf#11 @NoraLoose
Missing requirements-dev.txt referenced in contributing guide
On the contributing page https://pyturbo-sf.readthedocs.io/en/latest/contributing.html the installation instructions state:
```
# Install package in development mode
pip install -e .

# Install development dependencies
pip install -r requirements-dev.txt
```
However, there does not appear to be a `requirements-dev.txt` file in the main repository directory.

This issue is part of the JOSS review over https://github.com/openjournals/joss-reviews/issues/9876.
