### 2026-03-04 aayouche/pyturbo_sf#12 @NoraLoose
Figures on Quick Start page
In the [Quick Start Guide](https://pyturbo-sf.readthedocs.io/en/latest/quickstart.html), I noticed that the figures and print outputs from the Jupyter cells are not included in the rendered documentation.

Was this intentional? It could be helpful for users to see the expected figures and outputs so they can compare them with their own results. Of course, if there is a specific reason for omitting them (e.g., keeping the page concise), that’s completely fine. I'm just curious about the rationale! :)

This issue is part of the JOSS review over https://github.com/openjournals/joss-reviews/issues/9876.
