### 2026-03-04 aayouche/pyturbo_sf#13 @NoraLoose
Statement of need in documentation
As part of the JOSS review, I need to evaluate the following documentation criterion:

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Would you mind adding a short “Statement of Need” section to the documentation?

This issue is part of the JOSS review over https://github.com/openjournals/joss-reviews/issues/9876.
