### 2026-03-04 #9876 @NoraLoose
Hi @Sbozzolo and @aayouche, I read through the paper and tested the software/documentation. I think it’s a really cool package! 🎉

I’ve checked almost all the boxes in my review, but opened a few minor issues that I linked above (https://github.com/aayouche/pyturbo_sf/issues/4 - https://github.com/aayouche/pyturbo_sf/issues/15).
