### 2026-03-04 aayouche/pyturbo_sf#14 @NoraLoose
Error for "Multiple Orders" example in Quick Start Guide
When I try to run the https://pyturbo-sf.readthedocs.io/en/latest/quickstart.html#multiple-orders example, I get the error
```
KeyError                                  Traceback (most recent call last)
~/miniconda3/lib/python3.12/site-packages/xarray/core/dataset.py in ?(self, name)
   1237             variable = self._variables[name]
   1238         except KeyError:
-> 1239             _, name, variable = _get_virtual_variable(self._variables, name, self.sizes)
   1240

KeyError: 'sf_mean'

KeyError                                  Traceback (most recent call last)
File ~/miniconda3/lib/python3.12/site-packages/xarray/core/dataset.py:1344, in Dataset.__getitem__(self, key)
   1343 try:
-> 1344     return self._construct_dataarray(key)
   1345 except KeyError as e:

File ~/miniconda3/lib/python3.12/site-packages/xarray/core/dataset.py:1239, in Dataset._construct_dataarray(self, name)
   1238 except KeyError:
-> 1239     _, name, variable = _get_virtual_variable(self._variables, name, self.sizes)
   1241 needed_dims = set(variable.dims)

File ~/miniconda3/lib/python3.12/site-packages/xarray/core/dataset_utils.py:79, in _get_virtual_variable(variables, key, dim_sizes)
     78 if len(split_key) != 2:
---> 79     raise KeyError(key)
     81 ref_name, var_name = split_key

KeyError: 'sf_mean'

KeyError                                  Traceback (most recent call last)
/var/folders/ym/r6d93p_x5gz5cm14_qcmvsj00000gn/T/ipykernel_72372/4063845584.py in ?()
      9
     10 # Plot scaling
     11 plt.figure(figsize=(10, 8))
     12 for n in orders:
---> 13     r = results[n]['sf_mean'].coords['time']
     14     sf = results[n]['sf_mean'].values
     15     plt.loglog(r, sf, 'o-', label=f'Order {n}')
     16

~/miniconda3/lib/python3.12/site-packages/xarray/core/dataset.py in ?(self, key)
   1353
   1354                 # If someone attempts `ds['foo' , 'bar']` instead of `ds[['foo', 'bar']]`
   1355                 if isinstance(key, tuple):
   1356                     message += f"\nHint: use a list to select multiple variables, for example `ds[{list(key)}]`"
-> 1357                 raise KeyError(message) from e
   1358
   1359         if utils.iterable_of_hashable(key):
   1360             return self._copy_listed(key)

KeyError: "No variable named 'sf_mean'. Variables on the dataset include ['sf', 'std_error', 'ci_upper', 'ci_lower', 'nbootstraps', 'density', 'point_counts', 'converged', 'bin', 'time_bins']"
```

This issue is part of the JOSS review over on https://github.com/openjournals/joss-reviews/issues/9876.
