### 2026-03-04 aayouche/pyturbo_sf#15 @NoraLoose
Quick Start Guide multiple-variables example fails due to undefined signal/n
When I'm trying to run the https://pyturbo-sf.readthedocs.io/en/latest/quickstart.html#multiple-variables example in the Quick Start Guide, I get the error
```
Cell In[15], line 5
      1 # Create dataset with multiple variables
      2 ds_multi = xr.Dataset(
      3     data_vars={
      4         "velocity": ("time", signal),
----> 5         "temperature": ("time", signal + 0.5*np.random.randn(n))
      6     },
      7     coords={"time": time}
      8 )
     10 variables = ["velocity", "temperature"]
     11 results_multi = {}

ValueError: operands could not be broadcast together with shapes (5000,) (2,)
```

Looks like it's grabbing `n` and `signal` from previous examples in my notebook, which are not consistent. Maybe safer to redefine `signal` and `n` in this jupyter cell?

This issue is part of the JOSS review over on https://github.com/openjournals/joss-reviews/issues/9876.
