### 2026-03-06 aayouche/pyturbo_sf#11 @aayouche
this has been fixed in the doc

   pip install -e ".[dev]"
