### 2026-03-06 3c44caa
Update installation.rst
