### 2026-03-06 c56ce16
Update installation.rst
