### 2026-03-06 aayouche/pyturbo_sf#12 @aayouche
Thank you for the suggestion! The outputs were intentionally omitted to keep the documentation concise and reduce build complexity. Since structure function results depend on the random seed used for synthetic data generation, the exact numerical values would vary between runs anyway.
However, we agree that seeing expected output shapes and plot styles would help users. We'll consider adding static example figures in a future update, or converting key examples to executed notebooks using nbsphinx.
