### 2026-03-06 21d04ab
Update paper.md
