+++ paper/paper.md
@@ -33,7 +33,7 @@ affiliations:
    index: 1
  - name: Woods Hole Oceanographic Institution, Woods Hole, MA, USA
    index: 2
- - name: Department of Earth, Environmental, and Planetary Sciences (DEEPS), Brown University, Providence, RI, USA
+ - name: College of Earth, Ocean, and Atmospheric Sciences, Oregon State University, Corvallis, OR, USA
    index: 3
 
 date: 30 September 2025
@@ -42,61 +42,49 @@ bibliography: paper.bib
 
 # Summary
 
-Structure functions are fundamental diagnostic tools in turbulence research that quantify spatial correlations between field differences at varying separation distances, revealing energy cascade characteristics and scaling laws [@Frisch1995; @Pope2000]. PyTurbo_SF is a comprehensive Python package providing efficient, statistically rigorous structure function calculations for 1D, 2D, and 3D turbulent datasets through an innovative adaptive bootstrap framework.
+Structure functions (SFs) are fundamental diagnostic tools in turbulence research that quantify spatial correlations between field differences at varying separation distances, revealing energy cascade characteristics and scaling laws [@Frisch1995; @Pope2000]. PyTurbo_SF is a comprehensive Python package providing efficient, statistically rigorous structure function calculations for 1D, 2D, and 3D turbulent datasets through an innovative adaptive bootstrap framework.
 
 The package addresses computational bottlenecks through power-of-2 spacing strategies, adaptive convergence monitoring, and memory-optimized algorithms. PyTurbo_SF delivers robust uncertainty quantification through bootstrap resampling and supports diverse (conditional) structure function types including longitudinal, transverse, scalar, advective, and energy flux functions [@Pearson2021; @Pearson2024].
 
-
-Applications span oceanographic time series and satellite measurements to high-resolution simulations, enabling consistent methodology across scales from laboratory to planetary systems.
+Applications span oceanographic time series and satellite measurements to high-resolution simulations, enabling consistent methodology across scales from laboratory to planetary systems. Despite its name evoking the broader turbulence field, PyTurbo_SF is specifically designed for structure function analysis — one of the most widely used diagnostic approaches in turbulence research.
 
 # Statement of need
 
 Contemporary turbulence research relies on massive datasets from satellite missions, autonomous platforms, and high-resolution simulations. Traditional structure function calculations face severe limitations: computational intractability for large datasets, absence of uncertainty quantification, manual parameter tuning, and limited function types.
 
-Existing tools address only subsets of these challenges. fastSF provides parallelized implementations but lacks advanced function types and uncertainty quantification [@Sadhukhan2021]. MATLAB toolkits are environmentally limited and lack comprehensive statistical frameworks [@Fuchs2022]. Alternative approaches like FlowSieve cannot provide the scale-by-scale information structure functions uniquely deliver [@Storer2023].
+Existing tools address only subsets of these challenges. fastSF provides parallelized implementations but lacks advanced function types and uncertainty quantification [@Sadhukhan2021]. MATLAB toolkits are environmentally limited and lack comprehensive statistical frameworks [@Fuchs2022]. FluidSF [@Wagner2025] is a related Python package supporting core structure function types, which we compare in detail below. Alternative approaches like FlowSieve cannot provide the scale-by-scale information structure functions uniquely deliver [@Storer2023].
 
[…]
