+++ paper/paper.bib
@@ -106,15 +106,40 @@
   doi={10.5334/jors.148}
 }
 
+@article{Morrow2019,
+  title={Global observations of fine-scale ocean surface topography with the {Surface Water and Ocean Topography (SWOT)} mission},
+  author={Morrow, Rosemary and Fu, Lee-Lueng and Ardhuin, Fabrice and Benkiran, Mounir and Chapron, Bertrand and Cosme, Emmanuel and d'Ovidio, Francesco and Farrar, J. Thomas and Gille, Sarah T. and Lapeyre, Guillaume and others},
+  journal={Frontiers in Marine Science},
+  volume={6},
+  pages={232},
+  year={2019},
+  publisher={Frontiers Media SA},
+  doi={10.3389/fmars.2019.00232}
+}
 
-@article{Morrow2018,
-  title={Fine-scale altimetry and the future {SWOT} mission},
-  author={Morrow, Rosemary and Blurmstein, Denis and Dibarboure, Gérald},
-  journal={New Frontiers in Operational Oceanography},
-  pages={191--226},
-  year={2018},
-  publisher={GODAE OceanView},
-  doi={10.17125/gov2018.ch08}
+@article{Shchepetkin2005,
+  title={The regional oceanic modeling system ({ROMS}): a split-explicit, free-surface, topography-following-coordinate oceanic model},
+  author={Shchepetkin, Alexander F. and McWilliams, James C.},
+  journal={Ocean Modelling},
+  volume={9},
+  number={4},
+  pages={347--404},
+  year={2005},
+  publisher={Elsevier},
+  doi={10.1016/j.ocemod.2004.08.002},
+  note={{CROCO} is built upon the {ROMS} framework}
+}
+
+@article{Stevens2005,
+  title={Evaluation of large-eddy simulations via observations of nocturnal marine stratocumulus},
+  author={Stevens, Bjorn and Moeng, Chin-Hoh and Ackerman, Andrew S. and Bretherton, Christopher S. and Chlond, Andreas and de Roode, Stephan and Edwards, James and Golaz, Jean-Christophe and Jiang, Hongli and Khairoutdinov, Marat and others},
+  journal={Monthly Weather Review},
+  volume={133},
+  number={6},
+  pages={1443--1462},
+  year={2005},
+  publisher={American Meteorological Society},
+  doi={10.1175/MWR2930.1}
 }
 
 @article{Monin1975, 
