### 2026-03-06 4b1a1f0
Update structure_functions.rst
