### 2026-03-06 #9876 @aayouche
Dear @Sbozzolo , @einaraz , and @NoraLoose ,

Thank you for your thorough and constructive reviews of PyTurbo_SF. Your detailed feedback has significantly improved both the paper and the software documentation. We appreciate the time and effort you invested in evaluating our work, and we are pleased to report that we have addressed all of your comments and suggestions.

Below we provide point-by-point responses to each reviewer's comments, detailing the changes made.

Best regards,
Adam Ayouche and co-authors

---

## Response to @einaraz

### Comment 1: Package name and logo clarity
> The package name and initial logo suggest that it is intended for turbulence analysis, which is a very broad field. However, it is designed for structure function analysis, a specific type of analysis within turbulence. I wonder if this should be clarified earlier, including the description in the logo.

**Response:** We have addressed this in multiple ways:

1. **Paper clarification:** Added explicit statement in the Summary: *"Despite its name evoking the broader turbulence field, PyTurbo_SF is specifically designed for structure function analysis — one of the most widely used diagnostic approaches in turbulence research."*

2. **Logo update ** We have updated the logo.

3. **Documentation:** The index.rst and README.md now clearly state in the opening paragraph that PyTurbo_SF is for "structure function calculations."

---

### Comment 2: Eddy-covariance systems not mentioned
> Paper: one major source of turbulent data comes from field experiments using eddy-covariance systems, which is not mentioned. This package would be very useful for this community and an example could be easily added in the tutorial if the authors wanted to explore that area.

**Response:** We have added mentions of eddy-covariance applications in two places:

1. **Statement of Need:** *"Similarly, flux tower networks (e.g., AmeriFlux, FLUXNET) generate extensive eddy-covariance time series of velocity and scalar turbulence that would benefit from standardized structure function diagnostics with uncertainty quantification."*

2. **Software Functionality:** *"Applications extend beyond traditional oceanographic and atmospheric modeling contexts to include eddy-covariance field measurements, where high-frequency sonic anemometer time series of wind components and scalars can be analyzed using `bin_sf_1d()` for temporal structure functions."*

We agree this is an excellent application area and will consider adding a dedicated eddy-covariance example notebook in a future release.

---

### Comment 3: Figure captions need variable definitions
> Figures 1 and 2: Please define all variables in the figure captions. It would also help to include a short paragraph in the main text explaining the figures and specifying which functions were used to generate them and what raw data served as input.

**Response:** We have completely rewritten all figure captions (Figures 1–4) to include:

- Definitions of all variables with units
- Description of input data sources with citations
- Specification of which PyTurbo_SF functions were used (e.g., `get_isotropic_sf_2d()`, `bin_sf_2d()`, `bin_sf_3d()`)
- Function parameters used (e.g., `fun='longitudinal'`, `order=3`)
- Physical interpretation of results
- Explanation of uncertainty visualization (bootstrap confidence intervals)

See the revised paper for the detailed captions.

---

### Comment 4: Python version inconsistencies
> Installation: I successfully installed the package using pip install pyturbo_sf (tested on Python 3.9). However, installation from the repository appears to require Python 3.11 or higher. Additionally, the pyproject.toml and requirements.txt files inside the docs/ directory seem mismatched.

**Response:** We have standardized all Python version references to **Python 3.12+**:

| File | Change |
|------|--------|
| `pyproject.toml` | `requires-python = ">=3.12"`, removed Python 3.11 and 3.14 classifiers |
| `README.md` | Badge updated to `python-3.12%2B` |
| `index.rst` | Badge updated to `python-3.12%2B` |
| `installation.rst` | States "Python 3.12 or higher" |
| `contributing.rst` | Uses `python=3.12` in examples |
| `conf.py` | Version updated to `1.0.7` |

We also updated dependency version floors in `pyproject.toml` to realistic values compatible with Python 3.12.

---

### Comment 5: Documentation quality
> Documentation: The documentation is very detailed, and I was able to run several of the provided examples successfully.

**Response:** Thank you for this positive feedback! We have further improved the documentation by:

- Adding a "Statement of Need" section to `index.rst`
- Adding physical interpretations to all 11 example notebooks
- Implementing the `sf_kwargs` pattern in notebooks for cleaner, more compact code
- Fixing self-contained examples in the quickstart guide

---

## Response to @NoraLoose

### Figure 1 Comments
> The caption could be a bit more descriptive. For example, you could explain that the upper panel shows zonal velocity from SWOT along the satellite track (maybe with citation for SWOT to point non-expert readers to a reference), and the lower panel shows the structure function analysis. Maybe you could clarify what "longitudinal" means. Does it mean the separation vector points along longitude? ... You mention "bootstrap error bars," but I don't see any in the lower panel.

**Response:** The Figure 1 caption has been completely rewritten to address all points:

- Added SWOT citation [@Morrow2019]
- Explained the upper panel shows zonal geostrophic velocity from SWOT's twin Ka-band radar interferometer swaths
- Clarified X/Y are local Cartesian coordinates (eastward/northward distances in km)
- Explained "longitudinal" means the velocity component projected onto the separation vector direction ($\delta u_L = \delta \mathbf{u} \cdot \hat{\mathbf{r}}$) — this is now also defined in the Software Functionality section
- Specified functions used: `get_isotropic_sf_2d()` with `fun='longitudinal', order=3` and `fun='advective'`
- Clarified the shaded regions ARE the bootstrap confidence intervals (95%)
- Added physical interpretation: positive values = forward cascade, negative = inverse cascade
- Noted the sign change near 100 km indicates cascade regime transition

---

### Figure 2 Comments
> Maybe add a citation for CROCO. And you could describe the panels in the caption so readers can follow along.

**Response:** Figure 2 caption now includes:

- CROCO citation [@Shchepetkin2005] (CROCO is built on the ROMS framework)
- Systematic description of all four panels:
  - Upper left: Absolute vorticity ω = ζ_z + f
  - Lower left: Zonal velocity u
  - Upper right: Third-order and advective SFs (ASF_V, SF_{LLL}/r, -ASF_ω, -SF_{Lωω}/r) with units and axes
  - Lower right: Traditional second-order SFs (SF_{LL}, SF_{TT}, SF_{ωω}) with units
- Specified all quantities plotted with their units and which axis they use
- Noted shaded regions = 95% bootstrap confidence intervals

---

### Figure 3 Comments
> It would help to clarify the axes. What exactly do x and y represent? If the figure highlights anisotropic features, you could state that explicitly.

**Response:** Figure 3 caption now explains:

- Axes represent separation distances in x (zonal) and y (meridional) directions, from 10³ to 10⁵ m on logarithmic scales
- Explicitly states anisotropic features: "The SFs are not perfectly symmetric about the diagonal, indicating that turbulent energy transfer differs between zonal and meridional directions"
- Physical interpretation: anisotropy reflects influence of background flow and mesoscale eddy structures
- Notes white regions indicate insufficient data pairs for reliable statistics
- Specifies `bin_sf_2d()` function used

---

### Figure 4 Comments
> Maybe include a citation for DYCOMS. And what exactly does the figure show?

**Response:** Figure 4 caption now includes:

- DYCOMS-II citation [@Stevens2005]
- Clear explanation: shows second-order longitudinal SF (SF_{LL}, m²s⁻²) computed with `bin_sf_3d()`
- Clarified axes: r_x, r_y (horizontal, 0–3500 m) and r_z (vertical, 0–45000 m) — noting different scales
- Color scale range: ~0.3 to ~0.9 m²s⁻²
- Physical interpretation: horizontal isotropy (symmetric in r_x–r_y plane), vertical anisotropy from stable stratification at boundary layer top
- Notes this enables assessment of turbulence isotropy and identification of characteristic length scales

---

### Line 67: SF abbreviation
> Abbreviation SF for structure function has not been introduced.

**Response:** Fixed. The abbreviation "SF" is now introduced in the first paragraph of the Summary: *"Structure functions (SFs) are fundamental diagnostic tools..."*

---

### FluidSF mention in Statement of Need
> Maybe mention FluidSF already in the Statement of need when you discuss other related tools.

**Response:** Added FluidSF mention in Statement of Need: *"FluidSF [@Wagner2025] is a related Python package supporting core structure function types, which we compare in detail below."*
