+++ paper/paper.md
@@ -70,13 +70,13 @@ The package provides three main interfaces: `bin_sf_1d()` for time series, `bin_
 
 Figures \ref{fig:swot}–\ref{fig:dycoms} demonstrate PyTurbo_SF applied across scientific domains, illustrating the package's versatility for both satellite observations and numerical simulations.
 
-**Figure 1** presents an analysis of NASA's Surface Water and Ocean Topography (SWOT) satellite altimetry data [@Morrow2019] in the Gulf Stream region. The upper panel displays the zonal geostrophic velocity field ($u_g$, m/s) derived from sea surface height gradients, mapped onto a local Cartesian coordinate system where $X$ and $Y$ represent eastward and northward distances in kilometers. The two diagonal swaths correspond to SWOT's twin Ka-band radar interferometer tracks. The lower panel shows isotropically-averaged structure functions computed using `get_isotropic_sf_2d()`: the third-order longitudinal SF divided by separation distance ($\mathrm{SF}_{LLL}/r$, blue) computed with `fun='longitudinal'` and `order=3`, and the advective SF ($\mathrm{ASF}_V$, red) computed with `fun='advective'`. Both quantities have units of energy dissipation rate (m$^2$\,s$^{-3}$). The shaded regions indicate 95% bootstrap confidence intervals. Physically, these third-order SFs are proportional to the spectral energy flux: positive values indicate forward cascade (energy transfer to smaller scales) while negative values indicate inverse cascade (energy transfer to larger scales). The sign change near 100 km separation distance reveals the transition between cascade regimes characteristic of quasi-geostrophic ocean turbulence.
+![Analysis of NASA's Surface Water and Ocean Topography (SWOT) satellite altimetry data [@Morrow2019] in the Gulf Stream region. **Upper panel:** Zonal geostrophic velocity field ($u_g$, m/s) derived from sea surface height gradients, mapped onto a local Cartesian coordinate system where $X$ and $Y$ represent eastward and northward distances in kilometers; the two diagonal swaths correspond to SWOT's twin Ka-band radar interferometer tracks. **Lower panel:** Isotropically-averaged structure functions computed using `get_isotropic_sf_2d()`: the third-order longitudinal SF divided by separation distance ($\mathrm{SF}_{LLL}/r$, blue; `fun='longitudinal'`, `order=3`) and the advective SF ($\mathrm{ASF}_V$, red; `fun='advective'`), both in units of m$^2$\,s$^{-3}$. Shaded regions indicate 95% bootstrap confidence intervals. These third-order SFs are proportional to spectral energy flux: positive values indicate forward cascade (energy transfer to smaller scales) while negative values indicate inverse cascade (energy transfer to larger scales). The sign change near 100 km separation distance reveals the transition between cascade regimes characteristic of quasi-geostrophic ocean turbulence.\label{fig:swot}](figs/figure1.png)
 
[…]
