### 2026-03-06 #9876 @aayouche
> Thanks [@aayouche](https://github.com/aayouche) for working on the paper changes. I really like the new descriptive figure captions. They’re great!
>
> Just one minor comment:
>
>     * Captions of Figures 1–4: I think the comma in “m^2 s^-3” should be removed (currently written as “m^2, s^-3”).

@NoraLoose Thanks so much it was the case for other captions as well - should be fixed now
