### 2026-03-06 aayouche/pyturbo_sf#7 @aayouche
should be fixed now for all the notebooks and the docs - thanks again for pointing out this inconsistency
