### 2026-03-12 aayouche/pyturbo_sf#16 @NoraLoose
Add GitHub Actions CI for automated testing
The repo has many tests, but no CI is set up. I suggest adding a GitHub Actions workflow to:

* Install .[dev] dependencies
* Run tests with pytest
* Trigger on PRs and pushes

This will automatically catch errors and ensure code quality across supported Python versions.

This issue is part of the JOSS review over on https://github.com/openjournals/joss-reviews/issues/9876.
