### 2026-03-12 aayouche/pyturbo_sf#10 @NoraLoose
This is great, thanks for adding these!

I just noticed a formatting issue in the docs. Things like
```
`:math:`epsilon`
```
show up throughout the rendered documentation (within the physical interpretations).
