### 2026-03-13 4d52d87
Update example_1D_mooring.ipynb
