### 2026-03-13 aayouche/pyturbo_sf#10 @aayouche
Thanks - the issue has been resolved and also pyturbo is now replaced in the examples ( it was a build problem) :)
