### 2026-03-13 aayouche/pyturbo_sf#16 @aayouche
Thank you for the suggestion! A GitHub Actions CI workflow has been added at `.github/workflows/ci.yml`. It installs `.[dev]` dependencies via `pip install -e ".[dev]"`, runs the full test suite with `pytest` , and triggers on every push and pull request to `main`. The workflow runs across all supported Python versions (3.11, 3.12, and 3.13) in parallel.

The CI status and documentation build status are now visible directly from the repository homepage via badges
