### 2026-03-13 #9876 @NoraLoose
Hi @Sbozzolo, I’ve finished my review. @aayouche and I worked through all the issues I opened in the repo, and they’ve now all been resolved. I’m also happy with the paper and have checked all the boxes in the checklist. From my side, this submission looks ready to be published.

Nice work, @aayouche!
