### 2026-03-13 #9876 @Sbozzolo
@aayouche, thank you for your responses. I see that all the issues are closed. Please, let me know if you have any pending work before I can take a look at the paper/repo/responses.

Thank you!
