### 2026-03-13 #9876 @Sbozzolo
Thank you, @aayouche. The submission is in great shape.

Here's some comments/questions:
- Could please have a look at the missing DOI reported in one of the messages above?
- The paper is a little on the longer side, do you think you could make it a little terser?
- Many of the labels on the figures are very small and hard to read
- I would recommend citing the "basic" software package you are using, e.g., xarray, joblib, numpy, when you mention them in the paper.
