+++ paper/paper.md
@@ -46,7 +46,7 @@ Structure functions (SFs) are fundamental diagnostic tools in turbulence researc
 
 The package addresses computational bottlenecks through power-of-2 spacing strategies, adaptive convergence monitoring, and memory-optimized algorithms. PyTurbo_SF delivers robust uncertainty quantification through bootstrap resampling and supports diverse (conditional) structure function types including longitudinal, transverse, scalar, advective, and energy flux functions [@Pearson2021; @Pearson2024].
 
-Applications span oceanographic time series and satellite measurements to high-resolution simulations, enabling consistent methodology across scales from laboratory to planetary systems. Despite its name evoking the broader turbulence field, PyTurbo_SF is specifically designed for structure function analysis — one of the most widely used diagnostic approaches in turbulence research.
+Applications span oceanographic time series and satellite measurements to high-resolution simulations, enabling consistent methodology across scales from laboratory to planetary systems. PyTurbo_SF is specifically designed for structure function analysis — one of the most widely used diagnostic approaches in turbulence research.
 
 # Statement of need
 
@@ -54,7 +54,7 @@ Contemporary turbulence research relies on massive datasets from satellite missi
 
 Existing tools address only subsets of these challenges. fastSF provides parallelized implementations but lacks advanced function types and uncertainty quantification [@Sadhukhan2021]. MATLAB toolkits are environmentally limited and lack comprehensive statistical frameworks [@Fuchs2022]. FluidSF [@Wagner2025] is a related Python package supporting core structure function types, which we compare in detail below. Alternative approaches like FlowSieve cannot provide the scale-by-scale information structure functions uniquely deliver [@Storer2023].
 
-There is a growing need for tools to analyze emerging datasets from NASA's SWOT satellite [@Morrow2019] and next-generation atmospheric simulations generating terabyte-scale outputs. Similarly, flux tower networks (e.g., AmeriFlux, FLUXNET) generate extensive eddy-covariance time series of velocity and scalar turbulence that would benefit from standardized structure function diagnostics with uncertainty quantification. Recent advances in structure function theory, particularly advective structure functions [@Pearson2021] and spectral flux estimation [@Pearson2024], require frameworks handling both traditional and novel function types with statistical rigor.
+There is a growing need for tools to analyze emerging datasets from NASA's SWOT satellite [@Morrow2019] and next-generation atmospheric simulations generating terabyte-scale outputs. Flux tower networks (e.g., AmeriFlux, FLUXNET) similarly generate extensive eddy-covariance time series that would benefit from standardized structure function diagnostics with uncertainty quantification. Recent advances in structure function theory, particularly advective structure functions [@Pearson2021] and spectral flux estimation [@Pearson2024], require frameworks handling both traditional and novel function types with statistical rigor.
 
 PyTurbo_SF fills this gap by providing the first comprehensive, statistically robust framework that scales from small observational datasets to massive simulation outputs while delivering quantified uncertainties essential for scientific interpretation.
 
@@ -62,9 +62,7 @@ PyTurbo_SF fills this gap by providing the first comprehensive, statistically ro
 
[…]
