+++ paper/paper.bib
@@ -43,9 +43,9 @@
 @article{Pearson2024,
   title={Estimating spectral fluxes in quasi-two-dimensional flows with advective structure functions and {Bessel} functions},
   author={Pearson, Brodie and Wagner, Cassidy and Fox-Kemper, Baylor and Samelson, Roger},
-  journal={Journal of Fluid Mechanics},
-  note={Under revision},
-  year={2024}
+  journal={Journal of Physical Oceanography},
+  year={2024},
+  doi={10.1175/jpo-d-24-0211.1}
 }
 
 @article{Sadhukhan2021,
@@ -186,6 +186,14 @@
   year={2017},
   publisher={Nature Publishing Group},
   doi={10.1038/nphys4227}
+}
+
+@misc{Joblib2020,
+  author = {{Joblib Development Team}},
+  title = {Joblib: running {Python} functions as pipeline jobs},
+  year = {2020},
+  url = {https://joblib.readthedocs.io/},
+  howpublished = {\url{https://joblib.readthedocs.io/}}
 }
 
 @article{Wagner2025,
