### 2026-03-14 #9876 @aayouche
Dear @Sbozzolo ,

Thank you for your comments. The DOI for Pearson et al. (2024) has been updated to 10.1175/jpo-d-24-0211.1 (Journal of Physical Oceanography). For joblib, no DOI exists for this package — we have included its documentation URL instead. We have also added explicit citations for NumPy, xarray, and joblib where they are mentioned in the text.

We have tightened the manuscript throughout by merging redundant paragraphs and streamlining the Related work section. The paper now stands at approximately 1,624 words — we apologize we could not reduce it further without removing content we feel is necessary to properly contextualize the package.

Figures 1, 2, and 3 have been redesigned at the correct JOSS print width (7 inches) with font sizes calibrated for the printed page. The original figures were designed at screen size and scaled down by the PDF renderer, which caused the small labels. Regarding Figure 4, the 3D volumetric plot with three separation-distance axes imposes inherent constraints on label size — we have done our best but cannot improve it further without changing the plot type entirely.
