### 2026-03-16 #9876 @Sbozzolo
Thank you, @aayouche.

I think that the labels are still too small to read comfortably (in particular, sub and superscripts and colored text). Could you re-generate your plots with larger labels?

There's also some visual artifact that hampers readability.

> Regarding Figure 4, the 3D volumetric plot with three separation-distance axes imposes inherent constraints on label size — we have done our best but cannot improve it further without changing the plot type entirely.

I don't see what you mean with this.

> JOSS print width (7 inches)

Where did you find this guideline? I don't think we have such requirement.
