### 2026-03-16 6af77ae
Update paper.md
