### 2026-03-17 #9876 @aayouche
Thanks @Sbozzolo .

Figures 1, 2, and 3 have been updated with larger fonts and figure sizes. Regarding Figure 4, our data is on NCAR's Derecho HPC which has been slow and intermittent — we've removed it for now and added a note pointing to the documentation examples. Happy to reinstate it later or leave it out if the remaining figures are sufficient.
