+++ paper/paper.md
@@ -68,7 +68,7 @@ Performance optimization enables analysis of previously intractable datasets. Me
 
 The package provides three main interfaces: `bin_sf_1d()` for time series, `bin_sf_2d()` for surface fields, and `bin_sf_3d()` for volumetric data. All functions automatically optimize computational strategies while supporting both isotropic and directional analysis. Applications extend beyond traditional oceanographic and atmospheric modeling contexts to include eddy-covariance field measurements, where high-frequency sonic anemometer time series of wind components and scalars can be analyzed using `bin_sf_1d()` for temporal structure functions.
 
-Figures \ref{fig:swot}–\ref{fig:croco2} demonstrate PyTurbo_SF applied across scientific domains, illustrating the package's versatility for both satellite observations and numerical simulations. Three-dimensional structure function analysis using `bin_sf_3d()` is demonstrated in the package documentation examples.
+Figures \ref{fig:swot}–\ref{fig:dycoms} demonstrate PyTurbo_SF applied across scientific domains, illustrating the package's versatility for both satellite observations and numerical simulations.
 
 ![Analysis of NASA's Surface Water and Ocean Topography (SWOT) satellite altimetry data [@Morrow2019] in the Gulf Stream region. **Upper panel:** Zonal geostrophic velocity field ($u_g$, m/s) derived from sea surface height gradients, mapped onto a local Cartesian coordinate system where $X$ and $Y$ represent eastward and northward distances in kilometers; the two diagonal swaths correspond to SWOT's twin Ka-band radar interferometer tracks. **Lower panel:** Isotropically-averaged structure functions computed using `get_isotropic_sf_2d()`: the third-order longitudinal SF divided by separation distance ($\mathrm{SF}_{LLL}/r$, blue; `fun='longitudinal'`, `order=3`) and the advective SF ($\mathrm{ASF}_V$, red; `fun='advective'`), both in units of m$^2$\ s$^{-3}$. Shaded regions indicate 95% bootstrap confidence intervals. These third-order SFs are proportional to spectral energy flux: positive values indicate forward cascade (energy transfer to smaller scales) while negative values indicate inverse cascade (energy transfer to larger scales). The sign change near 100 km separation distance reveals the transition between cascade regimes characteristic of quasi-geostrophic ocean turbulence.\label{fig:swot}](figs/figure1.png)
 
@@ -76,6 +76,7 @@ Figures \ref{fig:swot}–\ref{fig:croco2} demonstrate PyTurbo_SF applied across 
 
 ![Two-dimensional (non-isotropic) structure functions from CROCO simulation computed using `bin_sf_2d()`, showing SF values as functions of separation distance in both $x$ (zonal) and $y$ (meridional) directions on logarithmic axes from $10^3$ to $10^5$ m. **Left:** Advective SF ($\mathrm{ASF}_V$, m$^2$\ s$^{-3}$). **Center:** Third-order longitudinal SF ($\mathrm{SF}_{LLL}$, m$^3$\ s$^{-3}$). **Right:** Second-order longitudinal SF ($\mathrm{SF}_{LL}$, m$^2$\ s$^{-2}$). Anisotropic features are evident: the SFs are not symmetric about the diagonal, indicating different turbulent energy transfer rates in zonal versus meridional directions, reflecting the influence of background flow and mesoscale eddy structures. White regions at large separations indicate insufficient data pairs for reliable statistics.\label{fig:croco2}](figs/figure3.png)
 
[…]
