### 2026-03-21 #9876 @Sbozzolo
> Thanks [@Sbozzolo](https://github.com/Sbozzolo) for your suggestions, the figures have been updated and everything should be resolved - Thanks again.

Thank you! I'll try to get back to you as soon as possible, but I am on a business trip until the end of next week, which makes things harder.
