+++ paper/paper.md
@@ -68,21 +68,15 @@ Performance optimization enables analysis of previously intractable datasets. Me
 
 The package provides three main interfaces: `bin_sf_1d()` for time series, `bin_sf_2d()` for surface fields, and `bin_sf_3d()` for volumetric data. All functions automatically optimize computational strategies while supporting both isotropic and directional analysis. Applications extend beyond traditional oceanographic and atmospheric modeling contexts to include eddy-covariance field measurements, where high-frequency sonic anemometer time series of wind components and scalars can be analyzed using `bin_sf_1d()` for temporal structure functions.
 
-Figure \ref{fig:swot} demonstrates PyTurbo_SF applied to NASA SWOT satellite altimetry in the Gulf Stream region. The third-order longitudinal and advective SFs reveal a sign change near 100 km separation, marking the transition from inverse to forward energy cascade characteristic of quasi-geostrophic ocean turbulence.
+Figures \ref{fig:swot}–\ref{fig:dycoms} demonstrate PyTurbo_SF applied across scientific domains, illustrating the package's versatility for both satellite observations and numerical simulations.
 
-![SWOT altimetry analysis in the Gulf Stream. **Upper:** zonal geostrophic velocity field ($u_g$, m/s). **Lower:** isotropic $\mathrm{SF}_{LLL}/r$ (blue) and advective SF $\mathrm{ASF}_V$ (red) in m$^2$ s$^{-3}$, with 95% bootstrap confidence intervals (shading). Positive values indicate forward cascade; negative values indicate inverse cascade.\label{fig:swot}](figs/figure1.png)
+![Analysis of NASA's Surface Water and Ocean Topography (SWOT) satellite altimetry data [@Morrow2019] in the Gulf Stream region. **Upper panel:** Zonal geostrophic velocity field ($u_g$, m/s) derived from sea surface height gradients, mapped onto a local Cartesian coordinate system where $X$ and $Y$ represent eastward and northward distances in kilometers; the two diagonal swaths correspond to SWOT's twin Ka-band radar interferometer tracks. **Lower panel:** Isotropically-averaged structure functions computed using `get_isotropic_sf_2d()`: the third-order longitudinal SF divided by separation distance ($\mathrm{SF}_{LLL}/r$, blue; `fun='longitudinal'`, `order=3`) and the advective SF ($\mathrm{ASF}_V$, red; `fun='advective'`), both in units of m$^2$\ s$^{-3}$. Shaded regions indicate 95% bootstrap confidence intervals. These third-order SFs are proportional to spectral energy flux: positive values indicate forward cascade (energy transfer to smaller scales) while negative values indicate inverse cascade (energy transfer to larger scales). The sign change near 100 km separation distance reveals the transition between cascade regimes characteristic of quasi-geostrophic ocean turbulence.\label{fig:swot}](figs/figure1.png)
 
-Figure \ref{fig:croco} applies the same framework to CROCO regional ocean model output [@Shchepetkin2005]. Both third-order energy-flux SFs and traditional second-order SFs are computed isotropically using `get_isotropic_sf_2d()`, with shaded 95% bootstrap confidence intervals on all quantities.
[…]
