### 2026-03-31 347efa5
bump version to 1.0.8
