+++ CITATION.cff
@@ -3,10 +3,19 @@ message: "If you use this software, please cite it as below."
 authors:
 - family-names: "Ayouche"
   given-names: "Adam"
-  orcid: "https://orcid.org/0000-0002-5648-4521"
+  orcid: "https://orcid.org/0009-0001-9075-5911"
 - family-names: "Fox-Kemper"
   given-names: "Baylor"
   orcid: "https://orcid.org/0000-0002-2871-2048"
+- family-names: "Hell"
+  given-names: "Momme"
+  orcid: "https://orcid.org/0000-0002-5754-3925"
+- family-names: "Pearson"
+  given-names: "Brodie"
+  orcid: "https://orcid.org/0000-0002-0202-0481"
+- family-names: "Wagner"
+  given-names: "Cassidy"
+  orcid: "https://orcid.org/0000-0002-1186-2082"
 title: "pyturbo_sf"
 version: 1.0.8
 doi: 10.5281/zenodo.10463863
