+++ CITATION.cff
@@ -18,6 +18,6 @@ authors:
   orcid: "https://orcid.org/0000-0002-1186-2082"
 title: "pyturbo_sf"
 version: 1.0.8
-doi: 10.5281/zenodo.10463863
-date-released: 2025-04-10
+doi: 10.5281/zenodo.19342650
+date-released: 2026-03-30
 url: "https://github.com/aayouche/pyturbo_sf"
