### 2026-03-31 #9876 @aayouche
Thanks @Sbozzolo i have finished checking the tasks everything looks fine - Thanks again for everything :)

- latest version is 1.0.8
- zenodo archive : https://doi.org/10.5281/zenodo.19342650
