### 2026-03-31 #9876 @Sbozzolo
@aayouche

- We recommend that the title and author list for the metadata in your Zenodo software archive exactly match those in your JOSS paper to create a cohesive archive, though this is not a strict requirement.
- Your author list on Zenodo is missing the affiliations
