### 2026-04-01 264420f
Update paper.md
remove last two paragraphs
