### 2026-04-09 1a495b4
Update structure_functions.rst
added advective and pressure work in the 3d table
